+++ extensions/lorem-ipsum/src/utils.tsx
@@ -78,11 +78,11 @@ export const showError = async (msg: string) => {
 };
 
 export const produceOutput = async (content: string) => {
-  const { action: preference = "clipboard" } = getPreferenceValues();
+  const { action } = getPreferenceValues();
 
   await closeMainWindow();
 
-  switch (preference) {
+  switch (action) {
     case "clipboard":
       await Clipboard.copy(content);
       await showToast(Toast.Style.Success, "Copied to clipboard! 📋");
@@ -91,6 +91,12 @@ export const produceOutput = async (content: string) => {
     case "paste":
       await Clipboard.paste(content);
       await showToast(Toast.Style.Success, "Pasted to active app! 📝");
+      break;
+
+    case "pasteAndCopy":
+      await Clipboard.paste(content);
+      await Clipboard.copy(content);
+      await showToast(Toast.Style.Success, "Pasted to active app and copied to clipboard! 📋");
       break;
   }
 };